pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sysml-builder"
version = "0.1.0"
description = "Requirement Contract centered SysML v2 transformer for the GFSE strict suite."
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
sysml-builder = "sysml_builder.cli:main"
sysml-suite = "sysml_builder.suite:main"
sysml-builder-sidecar = "sysml_builder.sidecar_cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
